// frontend-gym/src/app/model/member.model.ts
export class MemberModel {
    id?: any;
    name?: string;
    username?: string;
    password?: string;
    phone?: string;
    nic?: string;
    age?: any;
    gender?: string;
    height?: any;
    weight?: any;
    bmi?: any;
    workout_time?: any;
    workout_experience?: any;
    fitness_goal?: string;
    schedule?: string;


}